+++ examples/cat_facts/Android/shared/src/main/java/com/redbadger/catfacts/shared_types/Msg.java
@@ -1,4 +1,4 @@
-package com.redbadger.crux_core.shared_types;
+package com.redbadger.catfacts.shared_types;
 
 
 public abstract class Msg {
