+++ examples/cat_facts/Android/shared/src/main/java/com/redbadger/catfacts/shared_types/RequestBody.java
@@ -1,19 +1,20 @@
-package com.redbadger.crux_core.shared_types;
+package com.redbadger.catfacts.shared_types;
 
 
-public abstract class ResponseBody {
+public abstract class RequestBody {
 
     abstract public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError;
 
-    public static ResponseBody deserialize(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
+    public static RequestBody deserialize(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
         int index = deserializer.deserialize_variant_index();
         switch (index) {
-            case 0: return Http.load(deserializer);
-            case 1: return Time.load(deserializer);
+            case 0: return Time.load(deserializer);
+            case 1: return Http.load(deserializer);
             case 2: return Platform.load(deserializer);
             case 3: return KVRead.load(deserializer);
             case 4: return KVWrite.load(deserializer);
-            default: throw new com.novi.serde.DeserializationError("Unknown variant index for ResponseBody: " + index);
+            case 5: return Render.load(deserializer);
+            default: throw new com.novi.serde.DeserializationError("Unknown variant index for RequestBody: " + index);
         }
     }
 
@@ -23,37 +24,75 @@ public abstract class ResponseBody {
         return serializer.get_bytes();
     }
 
-    public static ResponseBody bcsDeserialize(byte[] input) throws com.novi.serde.DeserializationError {
+    public static RequestBody bcsDeserialize(byte[] input) throws com.novi.serde.DeserializationError {
         if (input == null) {
              throw new com.novi.serde.DeserializationError("Cannot deserialize null array");
         }
         com.novi.serde.Deserializer deserializer = new com.novi.bcs.BcsDeserializer(input);
-        ResponseBody value = deserialize(deserializer);
+        RequestBody value = deserialize(deserializer);
         if (deserializer.get_buffer_offset() < input.length) {
              throw new com.novi.serde.DeserializationError("Some input bytes were not read");
         }
         return value;
     }
 
-    public static final class Http extends ResponseBody {
-        public final java.util.List<@com.novi.serde.Unsigned Byte> value;
+    public static final class Time extends RequestBody {
+        public Time() {
+        }
 
-        public Http(java.util.List<@com.novi.serde.Unsigned Byte> value) {
+        public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
+            serializer.increase_container_depth();
+            serializer.serialize_variant_index(0);
+            serializer.decrease_container_depth();
+        }
+
+        static Time load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
+            deserializer.increase_container_depth();
+            Builder builder = new Builder();
+            deserializer.decrease_container_depth();
+            return builder.build();
+        }
+
+        public boolean equals(Object obj) {
+            if (this == obj) return true;
+            if (obj == null) return false;
+            if (getClass() != obj.getClass()) return false;
+            Time other = (Time) obj;
+            return true;
+        }
+
+        public int hashCode() {
+            int value = 7;
+            return value;
+        }
+
+        public static final class Builder {
+            public Time build() {
+                return new Time(
+                );
+            }
+        }
+    }
+
+    public static final class Http extends RequestBody {
+        public final String value;
+
+        public Http(String value) {
             java.util.Objects.requireNonNull(value, "value must not be null");
             this.value = value;
         }
 
         public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
             serializer.increase_container_depth();
-            serializer.serialize_variant_index(0);
-            TraitHelpers.serialize_vector_u8(value, serializer);
+            serializer.serialize_variant_index(1);
+            serializer.serialize_str(value);
             serializer.decrease_container_depth();
         }
 
         static Http load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
             deserializer.increase_container_depth();
             Builder builder = new Builder();
-            builder.value = TraitHelpers.deserialize_vector_u8(deserializer);
+            builder.value = deserializer.deserialize_str();
             deserializer.decrease_container_depth();
             return builder.build();
         }
@@ -74,7 +113,7 @@ public abstract class ResponseBody {
         }
 
         public static final class Builder {
-            public java.util.List<@com.novi.serde.Unsigned Byte> value;
+            public String value;
 
             public Http build() {
                 return new Http(
@@ -84,74 +123,19 @@ public abstract class ResponseBody {
         }
     }
 
-    public static final class Time extends ResponseBody {
-        public final String value;
-
-        public Time(String value) {
-            java.util.Objects.requireNonNull(value, "value must not be null");
-            this.value = value;
-        }
-
-        public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
-            serializer.increase_container_depth();
-            serializer.serialize_variant_index(1);
-            serializer.serialize_str(value);
-            serializer.decrease_container_depth();
-        }
-
-        static Time load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
-            deserializer.increase_container_depth();
-            Builder builder = new Builder();
-            builder.value = deserializer.deserialize_str();
-            deserializer.decrease_container_depth();
-            return builder.build();
-        }
-
-        public boolean equals(Object obj) {
-            if (this == obj) return true;
-            if (obj == null) return false;
-            if (getClass() != obj.getClass()) return false;
-            Time other = (Time) obj;
-            if (!java.util.Objects.equals(this.value, other.value)) { return false; }
-            return true;
-        }
-
-        public int hashCode() {
-            int value = 7;
-            value = 31 * value + (this.value != null ? this.value.hashCode() : 0);
-            return value;
-        }
-
-        public static final class Builder {
-            public String value;
-
-            public Time build() {
-                return new Time(
-                    value
-                );
-            }
-        }
-    }
-
-    public static final class Platform extends ResponseBody {
-        public final String value;
-
-        public Platform(String value) {
-            java.util.Objects.requireNonNull(value, "value must not be null");
-            this.value = value;
+    public static final class Platform extends RequestBody {
+        public Platform() {
         }
 
         public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
             serializer.increase_container_depth();
             serializer.serialize_variant_index(2);
-            serializer.serialize_str(value);
             serializer.decrease_container_depth();
         }
 
         static Platform load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
             deserializer.increase_container_depth();
             Builder builder = new Builder();
-            builder.value = deserializer.deserialize_str();
             deserializer.decrease_container_depth();
             return builder.build();
         }
@@ -161,31 +145,26 @@ public abstract class ResponseBody {
             if (obj == null) return false;
             if (getClass() != obj.getClass()) return false;
             Platform other = (Platform) obj;
-            if (!java.util.Objects.equals(this.value, other.value)) { return false; }
             return true;
         }
 
         public int hashCode() {
             int value = 7;
-            value = 31 * value + (this.value != null ? this.value.hashCode() : 0);
             return value;
         }
 
         public static final class Builder {
-            public String value;
-
             public Platform build() {
                 return new Platform(
-                    value
                 );
             }
         }
     }
 
-    public static final class KVRead extends ResponseBody {
-        public final java.util.Optional<java.util.List<@com.novi.serde.Unsigned Byte>> value;
+    public static final class KVRead extends RequestBody {
+        public final String value;
 
-        public KVRead(java.util.Optional<java.util.List<@com.novi.serde.Unsigned Byte>> value) {
+        public KVRead(String value) {
             java.util.Objects.requireNonNull(value, "value must not be null");
             this.value = value;
         }
@@ -193,14 +172,14 @@ public abstract class ResponseBody {
         public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
             serializer.increase_container_depth();
             serializer.serialize_variant_index(3);
-            TraitHelpers.serialize_option_vector_u8(value, serializer);
+            serializer.serialize_str(value);
             serializer.decrease_container_depth();
         }
 
         static KVRead load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
             deserializer.increase_container_depth();
             Builder builder = new Builder();
-            builder.value = TraitHelpers.deserialize_option_vector_u8(deserializer);
+            builder.value = deserializer.deserialize_str();
             deserializer.decrease_container_depth();
             return builder.build();
         }
@@ -221,7 +200,7 @@ public abstract class ResponseBody {
         }
 
         public static final class Builder {
-            public java.util.Optional<java.util.List<@com.novi.serde.Unsigned Byte>> value;
+            public String value;
 
             public KVRead build() {
                 return new KVRead(
@@ -231,25 +210,30 @@ public abstract class ResponseBody {
         }
     }
 
-    public static final class KVWrite extends ResponseBody {
-        public final Boolean value;
+    public static final class KVWrite extends RequestBody {
+        public final String field0;
+        public final java.util.List<@com.novi.serde.Unsigned Byte> field1;
 
-        public KVWrite(Boolean value) {
-            java.util.Objects.requireNonNull(value, "value must not be null");
-            this.value = value;
+        public KVWrite(String field0, java.util.List<@com.novi.serde.Unsigned Byte> field1) {
+            java.util.Objects.requireNonNull(field0, "field0 must not be null");
+            java.util.Objects.requireNonNull(field1, "field1 must not be null");
+            this.field0 = field0;
+            this.field1 = field1;
         }
 
         public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
             serializer.increase_container_depth();
             serializer.serialize_variant_index(4);
-            serializer.serialize_bool(value);
+            serializer.serialize_str(field0);
+            TraitHelpers.serialize_vector_u8(field1, serializer);
             serializer.decrease_container_depth();
         }
 
         static KVWrite load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
             deserializer.increase_container_depth();
             Builder builder = new Builder();
-            builder.value = deserializer.deserialize_bool();
+            builder.field0 = deserializer.deserialize_str();
+            builder.field1 = TraitHelpers.deserialize_vector_u8(deserializer);
             deserializer.decrease_container_depth();
             return builder.build();
         }
@@ -259,22 +243,64 @@ public abstract class ResponseBody {
             if (obj == null) return false;
             if (getClass() != obj.getClass()) return false;
             KVWrite other = (KVWrite) obj;
-            if (!java.util.Objects.equals(this.value, other.value)) { return false; }
+            if (!java.util.Objects.equals(this.field0, other.field0)) { return false; }
+            if (!java.util.Objects.equals(this.field1, other.field1)) { return false; }
             return true;
         }
 
         public int hashCode() {
             int value = 7;
-            value = 31 * value + (this.value != null ? this.value.hashCode() : 0);
+            value = 31 * value + (this.field0 != null ? this.field0.hashCode() : 0);
+            value = 31 * value + (this.field1 != null ? this.field1.hashCode() : 0);
             return value;
         }
 
         public static final class Builder {
-            public Boolean value;
+            public String field0;
+            public java.util.List<@com.novi.serde.Unsigned Byte> field1;
 
             public KVWrite build() {
                 return new KVWrite(
-                    value
+                    field0,
+                    field1
+                );
+            }
+        }
+    }
+
+    public static final class Render extends RequestBody {
+        public Render() {
+        }
+
+        public void serialize(com.novi.serde.Serializer serializer) throws com.novi.serde.SerializationError {
+            serializer.increase_container_depth();
+            serializer.serialize_variant_index(5);
+            serializer.decrease_container_depth();
+        }
+
+        static Render load(com.novi.serde.Deserializer deserializer) throws com.novi.serde.DeserializationError {
+            deserializer.increase_container_depth();
+            Builder builder = new Builder();
+            deserializer.decrease_container_depth();
+            return builder.build();
+        }
+
+        public boolean equals(Object obj) {
+            if (this == obj) return true;
+            if (obj == null) return false;
+            if (getClass() != obj.getClass()) return false;
+            Render other = (Render) obj;
+            return true;
+        }
+
+        public int hashCode() {
+            int value = 7;
+            return value;
+        }
+
+        public static final class Builder {
+            public Render build() {
+                return new Render(
                 );
             }
         }
